+++ Near/app/src/main/java/com/alarmy/near/presentation/feature/myprofile/components/NearSocialLoginBadge.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.alarmy.near.presentation.feature.myprofile.LoginType
+import com.alarmy.near.presentation.feature.myprofile.model.LoginType
 import com.alarmy.near.presentation.ui.theme.NearTheme
 
 @Composable
