+++ Near/app/src/main/java/com/alarmy/near/presentation/feature/myprofile/model/LoginType.kt
@@ -1,4 +1,4 @@
-package com.alarmy.near.presentation.feature.myprofile
+package com.alarmy.near.presentation.feature.myprofile.model
 
 import androidx.annotation.StringRes
 import com.alarmy.near.R
